?? Box the stream ??
no idea why I need to move anything to the heap at this point.
Everything is in the same function. But futures:: Api wants me
to give something that implements the "Unpin" trait; whatever that is.

Guess we allocating on the heap now :shrug:

Other people are confused about this too. See https://github.com/rust-lang/futures-rs/issues/2200